feat(cmd): add update subcommand
Added a new subcommand `update` to the `cmd` module. This subcommand allows updating repositories with the latest releases or tags from GitHub. The `update` subcommand fetches the latest release or tag for each repository specified in the configuration file and updates the revision accordingly. This enhancement improves the functionality of the command-line tool by providing the ability to easily update repositories to the latest versions.